serverfiles/requirements.txt
# core deps
google-generativeai
ngrok
flask_ngrok
numpy==1.22.0
cython==3.0.11
scipy==1.11.4
torch @ https://download.pytorch.org/whl/cu121_full/torch-2.5.0%2Bcu121-cp310-cp310-linux_x86_64.whl
torchaudio @ https://download.pytorch.org/whl/cu121_full/torchaudio-2.5.0%2Bcu121-cp310-cp310-linux_x86_64.whl
soundfile==0.12.1
librosa==0.10.0
scikit-learn==1.5.2
numba==0.60.0
inflect==7.4.0
tqdm==4.66.5
anyascii==0.3.2
PyYAML==6.0.2
fsspec==2024.6.1
aiohttp==3.10.10
packaging==24.1
pandas==1.5.3
mutagen==1.47.0
# deps for examples
Flask==2.2.5
# deps for inference
pysbd==0.3.4
# deps for notebooks
umap-learn==0.5.6
# deps for training
matplotlib==3.7.1
# coqui stack
trainer==0.0.36
# config management
coqpit==0.0.17
# chinese g2p deps
jieba==0.42.1
pypinyin==0.53.0
# korean
hangul-romanize==0.1.0
# gruut+supported langs
gruut[de,es,fr]==2.2.3
# deps for korean
jamo==0.4.1
nltk==3.8.1
g2pkk==0.1.2
# deps for bangla
bangla==0.0.2
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
#deps for tortoise
einops>=0.8.0
transformers==4.44.2
#deps for bark
encodec==0.1.1
# deps for XTTS
Unidecode==1.3.8
num2words==0.5.13
spacy==3.7.5